update
Done

- 전체 게시물 조회

- 자세한 게시판(조회수++ 기능 완료)

- 답변 달기(답변 달면 답변 수 ++ 기능 완료)

- 게시물 삭제 완료

- 정환이가 원하는 대로 이름 바꿨다

to do

- 조회수 순으로 출력

- 태그별 출력

- 제목 검색

- 작성자 검색

- 제목 + 작성자 검색

to do(진짜 나중이지만 생각해봐야 할 것)

- seqNum(고유 식별자) hash값으로 만들기

- cache 기능 구현

+++ dev/back-end/devstudent/src/main/java/com/hours22/devstudent/Command/Mutation.java
@@ -1,6 +1,7 @@
 package com.hours22.devstudent.Command;
 
 import com.coxautodev.graphql.tools.GraphQLMutationResolver;
+import com.hours22.devstudent.Entity.Answer;
 import com.hours22.devstudent.Entity.Board;
 import com.hours22.devstudent.Entity.SequenceID;
 import com.hours22.devstudent.Repository.BoardRepository;
@@ -25,20 +26,50 @@ public class Mutation implements GraphQLMutationResolver {
             sequenceIDRepository.save(sequenceID);
         }
         //SequenceID sequenceID = sequenceIDRepository.findTopBy_idAfterOrderBySeqNumDesc("Board");
-        SequenceID sequenceID = sequenceIDRepository.findTopBy_id("Board");
+        SequenceID sequenceID = sequenceIDRepository.findBy_id("Board");
         int seqNum = sequenceID.getSeqNum() + 1;
         sequenceID.setSeqNum(seqNum);
         System.out.println("SequenceID = " + sequenceID.toString());
         sequenceIDRepository.save(sequenceID);
-        Board board = new Board(String.valueOf(seqNum),title,author,tags,date,content);
+        String previews = (content.length() < 20) ? content : content.substring(0, 20);
+        Board board = new Board(String.valueOf(seqNum),title,author,tags,date,content, previews);
         boardRepository.save(board);
         System.out.println(board.toString());
         return board;
     }
+    public Answer createAnswer(String parentID, String author, String content, String date){
 
+        System.out.println("=== CreateAnswer ===");
+        if(sequenceIDRepository.countBy_id("Answer") == 0) {
+            System.out.println("Answer 없습니다.");
+            SequenceID sequenceID = new SequenceID("Answer", 0);
+            sequenceIDRepository.save(sequenceID);
+        }
+        if(boardRepository.countBy_id(parentID)==0){
+            Answer answer = new Answer("null","Exception","Not Exist Board","null");
+            return answer;
+        }
+        SequenceID sequenceID = sequenceIDRepository.findBy_id("Answer");
+        int seqNum = sequenceID.getSeqNum() + 1;
+        sequenceID.setSeqNum(seqNum);
+        System.out.println("SequenceID = " + sequenceID.toString());
+        sequenceIDRepository.save(sequenceID);
+        Answer answer = new Answer(String.valueOf(seqNum),author,content,date);
+        Board board = boardRepository.findQuestionBy_id(parentID);
+        List<Answer> answers = board.getAnswers();
+        answers.add(answer);
+        board.setAnswers(answers);
+        board.setAnswerCount(board.getAnswerCount()+1);
+        boardRepository.save(board);
+        return answer;
+    }
     public Board deleteBoard(String _id){
+        if(boardRepository.countBy_id(_id) == 0){
+            Board board = new Board("null","Exception","hours22",null,"null","Not Exist Board","Not Exist Board");
+            return board;
+        }
         System.out.println("=== Delete Board ===");
-        Board board = boardRepository.findBy_id(_id);
+        Board board = boardRepository.findQuestionBy_id(_id);
         boardRepository.delete(board);
         System.out.println(board.toString());
         return board;
